Step3 - 로또(2등) (#3325)
* refacotr: 통화 단위 타입 변경

- 소숫점 계산표기를 위한 BicDecimal 타입으로 변경

* refactor: 테스트를 위한 생성자 삭제

- 생성자 삭제 후 기능 테스트로 변경

* refactor: 테스트를 위한 생성자 삭제₩

* refactor: 결과 확인 메소드명 변경

* refacotr: 로또 등수 검색 기능 stream으로 변경

- 가동석과 뎁스를 줄이기 위해 for와 if의 조합에서 strem으로 변경

* feat: 로또 등수에 2등 추가

- 2등을 관리 하기 위한 boolean 타입의 변수 추가

* feat: 당첨번호와 일치하는 보너스 숫자가 있는지 확인 구현

- bonus 클래스를 생성하여 역할 위임

* fix: 2등 로또 상금 수정

* feat: resultView 2등 출력 구현

* refactor: lotto numbers -> lotto로 책임 위임

- 단순히 매개변수 전달 역할만 하고 있던 numbers에 대한 내용 위임

* remove: 사용하지 않는 numbers 클래스 삭제

- 클래스 삭제와 테스트 코드 이동

* fix: 당첨로또 생성자 유효성 검사 수정

* refactor: InputView와 ResultView static 제거

* refactor: Bonus를 LottoNumber로 교체

- 분석 결과 하는 역할이 같기 때문에 교체

* refactor: LottoRank 일치 수와 보너스 로직 변경

- 검증 로직 Lambda 식으로 변경
- 필요 없는 상수값 제거

* refactor: InputView 상수 접근 제어자 변경

* test: 로또 등수 구하기 test 추가 작성

- 모든 경우에 대한 작성

+++ src/test/java/lotto/domain/LottoPrizeTest.java
@@ -2,13 +2,15 @@ package lotto.domain;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 class LottoPrizeTest {
 
     @Test
     void 수량으로_금액을_계산_할_수_있다() {
-        LottoPrize prize = new LottoPrize(1000, Currency.KRW);
+        LottoPrize prize = new LottoPrize(BigDecimal.valueOf(1000), Currency.KRW);
 
         long actual = prize.calculateAmount(2);
         long expected = 2000;
